Add files via upload
I couldn't figure out how to return more than one variable from a function and keep them separate when referencing the result of the function, so I had to put print() in the function, thus making it a chatterbox function. I would love to hear your feedback on how to get it to work without the print() being inside the function.