Add files via upload
I did not get the same results as the book did, however, it might have to do with something in the difference in punctuation used. I had to mess with it quite a bit. I was also unsure as to if I had the right file because my results were different from the book's. I cannot seem to get my program to actually count accurately either. I do not know how to fix this. Please let me know if you know where I went wrong. Thank you